feat: Проверка прав пользователя

+++ app/controllers/channel_types_controller.ts
@@ -1,4 +1,6 @@
+import { accessErrorMessages } from '#helpers/access_error_messages'
 import ChannelType from '#models/channel_type'
+import ChannelTypePolicy from '#policies/channel_type_policy'
 import ChannelTypeService from '#services/channel_type_service'
 import { chanelTypeValidator } from '#validators/channel_type'
 import type { HttpContext } from '@adonisjs/core/http'
@@ -16,7 +18,11 @@ export default class ChannelTypesController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response, auth }: HttpContext) {
+  async store({ request, response, auth, bouncer }: HttpContext) {
+    if (await bouncer.with(ChannelTypePolicy).denies('create')) {
+      return response.status(403).json({ message: accessErrorMessages.create })
+    }
+
     const { user } = auth
     const validatedData = await request.validateUsing(chanelTypeValidator)
     const channelType = await ChannelType.create({ userId: user?.id, ...validatedData })
@@ -27,7 +33,11 @@ export default class ChannelTypesController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
+    if (await bouncer.with(ChannelTypePolicy).denies('edit')) {
+      return response.status(403).json({ message: accessErrorMessages.edit })
+    }
+
     const channelType = await ChannelType.findOrFail(params.id)
     const validatedData = await request.validateUsing(chanelTypeValidator)
     const updChannelType = await channelType.merge(validatedData).save()
@@ -38,7 +48,11 @@ export default class ChannelTypesController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
+    if (await bouncer.with(ChannelTypePolicy).denies('delete')) {
+      return response.status(403).json({ message: accessErrorMessages.delete })
+    }
+
     const channelType = await ChannelType.findOrFail(params.id)
 
     await channelType.delete()
