+++ app/controllers/completed_works_controller.ts
@@ -1,4 +1,6 @@
+import { accessErrorMessages } from '#helpers/access_error_messages'
 import CompletedWork from '#models/completed_work'
+import CompletedWorkPolicy from '#policies/completed_work_policy'
 import CompletedWorkService from '#services/completed_wokr_service'
 import { completedWorkValidator } from '#validators/completed_work'
 import type { HttpContext } from '@adonisjs/core/http'
@@ -16,7 +18,11 @@ export default class CompletedWorksController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response, auth }: HttpContext) {
+  async store({ request, response, auth, bouncer }: HttpContext) {
+    if (await bouncer.with(CompletedWorkPolicy).denies('create')) {
+      return response.status(403).json({ message: accessErrorMessages.create })
+    }
+
     const { user } = auth
     const validatedData = request.validateUsing(completedWorkValidator)
     const completedWork = await CompletedWork.create({ userId: user?.id, ...validatedData })
@@ -32,8 +38,13 @@ export default class CompletedWorksController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
     const completedWork = await CompletedWork.findOrFail(params.id)
+
+    if (await bouncer.with(CompletedWorkPolicy).denies('edit', completedWork)) {
+      return response.status(403).json({ message: accessErrorMessages.edit })
+    }
+
     const validatedData = await request.validateUsing(completedWorkValidator)
     const updCompletedWork = await completedWork.merge({ ...validatedData }).save()
 
@@ -43,7 +54,11 @@ export default class CompletedWorksController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
+    if (await bouncer.with(CompletedWorkPolicy).denies('delete')) {
+      return response.status(403).json({ message: accessErrorMessages.delete })
+    }
+
     const completedWork = await CompletedWork.findOrFail(params.id)
 
     await completedWork.delete()
