+++ app/controllers/heads_controller.ts
@@ -1,4 +1,6 @@
+import { accessErrorMessages } from '#helpers/access_error_messages'
 import HeadController from '#models/head_controller'
+import HeadControllerPolicy from '#policies/head_controller_policy'
 import { HeadControllersService } from '#services/head_controller_service'
 import { headControllerValidator } from '#validators/head_controller'
 import type { HttpContext } from '@adonisjs/core/http'
@@ -16,7 +18,11 @@ export default class HeadsController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response, auth }: HttpContext) {
+  async store({ request, response, auth, bouncer }: HttpContext) {
+    if (await bouncer.with(HeadControllerPolicy).denies('create')) {
+      return response.status(403).json({ message: accessErrorMessages.create })
+    }
+
     const { user } = auth
     const validatedData = await request.validateUsing(headControllerValidator)
     const newHeadController = await HeadController.create({ userId: user?.id, ...validatedData })
@@ -27,7 +33,11 @@ export default class HeadsController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
+    if (await bouncer.with(HeadControllerPolicy).denies('edit')) {
+      return response.status(403).json({ message: accessErrorMessages.edit })
+    }
+
     const headController = await HeadController.findOrFail(params.id)
     const validatedData = await request.validateUsing(headControllerValidator)
     const updHeadController = await headController.merge(validatedData).save()
@@ -38,7 +48,11 @@ export default class HeadsController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
+    if (await bouncer.with(HeadControllerPolicy).denies('delete')) {
+      return response.status(403).json({ message: accessErrorMessages.delete })
+    }
+
     const headController = await HeadController.findOrFail(params.id)
 
     await headController.delete()
