+++ app/controllers/types_kps_controller.ts
@@ -1,4 +1,6 @@
+import { accessErrorMessages } from '#helpers/access_error_messages'
 import TypeKp from '#models/type_kp'
+import TypeKpPolicy from '#policies/type_kp_policy'
 import TypeKpService from '#services/type_kp_service'
 import { typeKpValidator } from '#validators/type_kp'
 import type { HttpContext } from '@adonisjs/core/http'
@@ -16,7 +18,11 @@ export default class TypesKpsController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response, auth }: HttpContext) {
+  async store({ request, response, auth, bouncer }: HttpContext) {
+    if (await bouncer.with(TypeKpPolicy).denies('create')) {
+      return response.status(403).json({ message: accessErrorMessages.create })
+    }
+
     const { user } = auth
     const validatedData = await request.validateUsing(typeKpValidator)
     const typeKp = await TypeKp.create({ userId: user?.id, ...validatedData })
@@ -27,7 +33,11 @@ export default class TypesKpsController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
+    if (await bouncer.with(TypeKpPolicy).denies('edit')) {
+      return response.status(403).json({ message: accessErrorMessages.edit })
+    }
+
     const typeKp = await TypeKp.findOrFail(params.id)
     const validatedData = await request.validateUsing(typeKpValidator)
     const updTypeKp = await typeKp.merge(validatedData).save()
@@ -38,7 +48,11 @@ export default class TypesKpsController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
+    if (await bouncer.with(TypeKpPolicy).denies('delete')) {
+      return response.status(403).json({ message: accessErrorMessages.delete })
+    }
+
     const typeKp = await TypeKp.findOrFail(params.id)
 
     await typeKp.delete()
