+++ app/controllers/voltage_classes_controller.ts
@@ -1,4 +1,6 @@
+import { accessErrorMessages } from '#helpers/access_error_messages'
 import VoltageClass from '#models/voltage_class'
+import VoltageClassPolicy from '#policies/voltage_class_policy'
 import VoltageClassService from '#services/voltage_class_service'
 import { voltageClassValidator } from '#validators/voltage_class'
 import type { HttpContext } from '@adonisjs/core/http'
@@ -16,7 +18,11 @@ export default class VoltageClassesController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response, auth }: HttpContext) {
+  async store({ request, response, auth, bouncer }: HttpContext) {
+    if (await bouncer.with(VoltageClassPolicy).denies('create')) {
+      return response.status(403).json({ message: accessErrorMessages.create })
+    }
+
     const { user } = auth
     const validatedData = await request.validateUsing(voltageClassValidator)
     const voltageClass = await VoltageClass.create({ userId: user?.id, ...validatedData })
@@ -27,7 +33,11 @@ export default class VoltageClassesController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
+    if (await bouncer.with(VoltageClassPolicy).denies('edit')) {
+      return response.status(403).json({ message: accessErrorMessages.edit })
+    }
+
     const voltageClass = await VoltageClass.findOrFail(params.id)
     const validatedData = await request.validateUsing(voltageClassValidator)
     const updVoltageClass = await voltageClass.merge(validatedData).save()
@@ -38,7 +48,11 @@ export default class VoltageClassesController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
+    if (await bouncer.with(VoltageClassPolicy).denies('delete')) {
+      return response.status(403).json({ message: accessErrorMessages.delete })
+    }
+
     const voltageClass = await VoltageClass.findOrFail(params.id)
 
     await voltageClass.delete()
